Button Bindings
I have bound the buttons to their respective commands based on the whiteboard of button bindings in prog chat. Their are a few commands that aren't used and have to be figured out also I haven't done any of the align commands as those are done in drive not RobotContainer.